+++ tools/SIOWrapper.cpp
@@ -124,7 +124,7 @@ int SIOWrapper::SetSIOServerMode(ESIOServerCommandLine cmdLine)
 }
 
 int SIOWrapper::ReadSector(uint8_t driveNo, uint16_t sector, 
-		uint8_t* buf, size_t length,
+		uint8_t* buf, unsigned int length,
 		uint8_t highspeedMode)
 {
 	if (driveNo<1 || driveNo > 8 || buf==0 || length==0) {
@@ -146,7 +146,7 @@ int SIOWrapper::ReadSector(uint8_t driveNo, uint16_t sector,
 }
 
 int SIOWrapper::WriteSector(uint8_t driveNo, uint16_t sector, 
-		uint8_t* buf, size_t length,
+		uint8_t* buf, unsigned int length,
 		uint8_t highspeedMode)
 {
 	if (driveNo<1 || driveNo > 8 || buf==0 || length==0) {
@@ -168,7 +168,7 @@ int SIOWrapper::WriteSector(uint8_t driveNo, uint16_t sector,
 }
 
 int SIOWrapper::WriteAndVerifySector(uint8_t driveNo, uint16_t sector, 
-		uint8_t* buf, size_t length,
+		uint8_t* buf, unsigned int length,
 		uint8_t highspeedMode)
 {
 	if (driveNo<1 || driveNo > 8 || buf==0 || length==0) {
@@ -189,7 +189,7 @@ int SIOWrapper::WriteAndVerifySector(uint8_t driveNo, uint16_t sector,
 	return ExtSIO(params);
 }
 
-int SIOWrapper::FormatDisk(uint8_t driveNo, uint8_t* buf, size_t length,
+int SIOWrapper::FormatDisk(uint8_t driveNo, uint8_t* buf, unsigned int length,
 		uint8_t highspeedMode)
 {
 	if (driveNo<1 || driveNo > 8 || buf==0) {
@@ -465,7 +465,7 @@ int SIOWrapper::SendError()
 	return fLastResult;
 }
 
-int SIOWrapper::SendDataFrame(uint8_t* buf, size_t length)
+int SIOWrapper::SendDataFrame(uint8_t* buf, unsigned int length)
 {
 	SIO_data_frame frame;
 
@@ -483,7 +483,7 @@ int SIOWrapper::SendDataFrame(uint8_t* buf, size_t length)
 	return fLastResult;
 }
 
-int SIOWrapper::ReceiveDataFrame(uint8_t* buf, size_t length)
+int SIOWrapper::ReceiveDataFrame(uint8_t* buf, unsigned int length)
 {
 	SIO_data_frame frame;
 
@@ -501,7 +501,7 @@ int SIOWrapper::ReceiveDataFrame(uint8_t* buf, size_t length)
 	return fLastResult;
 }
 
-int SIOWrapper::SendRawFrame(uint8_t* buf, size_t length)
+int SIOWrapper::SendRawFrame(uint8_t* buf, unsigned int length)
 {
 	SIO_data_frame frame;
 
@@ -519,7 +519,7 @@ int SIOWrapper::SendRawFrame(uint8_t* buf, size_t length)
 	return fLastResult;
 }
 
-int SIOWrapper::ReceiveRawFrame(uint8_t* buf, size_t length)
+int SIOWrapper::ReceiveRawFrame(uint8_t* buf, unsigned int length)
 {
 	SIO_data_frame frame;
 
@@ -564,7 +564,7 @@ int SIOWrapper::SendCompleteXF551()
 	return fLastResult;
 }
 
-int SIOWrapper::SendDataFrameXF551(uint8_t* buf, size_t length)
+int SIOWrapper::SendDataFrameXF551(uint8_t* buf, unsigned int length)
 {
 	SIO_data_frame frame;
 
@@ -686,7 +686,7 @@ int SIOWrapper::SetTapeBaudrate(unsigned int baudrate)
 	return fLastResult;
 }
 
-int SIOWrapper::SendTapeBlock(uint8_t* buf, size_t length)
+int SIOWrapper::SendTapeBlock(uint8_t* buf, unsigned int length)
 {
 	SIO_data_frame frame;
